add: サブスク催促のUIや機能を共通のWidgetで切り出して実装してみました
