+++ lib/ui/screen/tutorial/tutorial_screen.dart
@@ -1,16 +1,14 @@
 import 'dart:async';
-import 'dart:ui';
 import 'package:confetti/confetti.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_sliding_tutorial/flutter_sliding_tutorial.dart';
 import 'package:food_gram_app/core/local/shared_preference.dart';
-import 'package:food_gram_app/core/theme/style/paywall_style.dart';
 import 'package:food_gram_app/core/theme/style/tutorial_style.dart';
 import 'package:food_gram_app/gen/assets.gen.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/router/router.dart';
-import 'package:food_gram_app/ui/screen/paywall/paywall_screen.dart';
+import 'package:food_gram_app/ui/component/paywall_widget.dart';
 import 'package:food_gram_app/ui/screen/setting/setting_view_model.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
@@ -105,143 +103,24 @@ class _TutorialScreenState extends ConsumerState<TutorialScreen> {
               ),
 
               /// 3ページ目
-              DecoratedBox(
-                decoration: BoxDecoration(
-                  image: DecorationImage(
-                    image: AssetImage(Assets.image.paywallBackground.path),
-                    fit: BoxFit.cover,
-                  ),
-                ),
-                child: BackdropFilter(
-                  filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
-                  child: ColoredBox(
-                    color: Colors.black.withValues(alpha: 0.4),
-                    child: SafeArea(
-                      child: Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 24),
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Container(
-                              padding: const EdgeInsets.all(20),
-                              decoration: BoxDecoration(
-                                color: Colors.white.withValues(alpha: 0.9),
-                                borderRadius: BorderRadius.circular(20),
-                                boxShadow: [
-                                  BoxShadow(
-                                    color: Colors.black.withValues(alpha: 0.1),
-                                    blurRadius: 10,
-                                    offset: const Offset(0, 5),
-                                  ),
-                                ],
-                              ),
-                              child: Column(
-                                children: [
-                                  Text(
-                                    l10n.paywallPremiumTitle,
-                                    style: TutorialStyle.title(),
-                                  ),
-                                  const Gap(16),
-                                  _buildFeatureItem(
-                                    icon: Icons.emoji_events,
-                                    title: l10n.paywallTrophyTitle,
-                                    description: l10n.paywallTrophyDesc,
-                                  ),
-                                  _buildFeatureItem(
-                                    icon: Icons.label,
-                                    title: l10n.paywallTagTitle,
-                                    description: l10n.paywallTagDesc,
-                                  ),
-                                  _buildFeatureItem(
-                                    icon: Icons.account_circle,
-                                    title: l10n.paywallIconTitle,
-                                    description: l10n.paywallIconDesc,
-                                  ),
-                                  _buildFeatureItem(
-                                    icon: Icons.block,
-                                    title: l10n.paywallAdTitle,
-                                    description: l10n.paywallAdDesc,
-                                  ),
-                                  Container(
-                                    padding: const EdgeInsets.all(14),
-                                    decoration: BoxDecoration(
-                                      gradient: LinearGradient(
-                                        colors: [
-                                          Colors.purple[400]!,
-                                          Colors.blue[400]!,
-                                        ],
-                                        begin: Alignment.topLeft,
-                                        end: Alignment.bottomRight,
-                                      ),
-                                      borderRadius: BorderRadius.circular(12),
-                                    ),
-                                    child: Row(
-                                      children: [
-                                        Container(
-                                          padding: const EdgeInsets.all(8),
-                                          decoration: BoxDecoration(
-                                            color: Colors.white
-                                                .withValues(alpha: 0.2),
-                                            borderRadius:
-                                                BorderRadius.circular(8),
-                                          ),
-                                          child: const Icon(
-                                            Icons.rocket_launch,
-                                            color: Colors.white,
-                                            size: 26,
-                                          ),
-                                        ),
-                                        const Gap(12),
-                                        Text(
-                                          l10n.paywallPrice,
-                                          style: const TextStyle(
-                                            fontSize: 20,
-                                            color: Colors.white,
-                                            fontWeight: FontWeight.bold,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                  const Gap(20),
-                                  SizedBox(
-                                    width: double.infinity,
-                                    height: 50,
-                                    child: ElevatedButton(
-                                      style: TutorialStyle.button(),
-                                      onPressed: () => _handlePurchase(l10n),
-                                      child: FittedBox(
-                                        child: Row(
-                                          children: [
-                                            Text(
-                                              l10n.paywallSubscribeButton,
-                                              style: TutorialStyle.close(),
-                                            ),
-                                          ],
-                                        ),
-                                      ),
-                                    ),
-                                  ),
-                                  const Gap(12),
-                                  TextButton(
-                                    onPressed: () async {
-                                      await pageController.nextPage(
-                                        duration:
-                                            const Duration(milliseconds: 500),
-                                        curve: Curves.easeInOut,
-                                      );
-                                    },
-                                    child: Text(
-                                      l10n.paywallSkip,
-                                      style: TutorialStyle.subTitle(),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ],
+              PaywallBackground(
+                child: SafeArea(
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 24),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        PaywallContent(
+                          onPurchase: _handlePurchase,
+                          onSkip: () async {
+                            await pageController.nextPage(
+                              duration: const Duration(milliseconds: 500),
+                              curve: Curves.easeInOut,
+                            );
+                          },
+                          showSkipButton: true,
                         ),
-                      ),
+                      ],
                     ),
                   ),
                 ),
@@ -348,7 +227,6 @@ class _TutorialScreenState extends ConsumerState<TutorialScreen> {
                         curve: Curves.easeInOut,
                       );
                     }
-                    // 最後のページ（4枚目）では何もしない
                   },
                 ),
               ),
@@ -360,139 +238,21 @@ class _TutorialScreenState extends ConsumerState<TutorialScreen> {
     );
   }
 
-  Widget _buildFeatureItem({
-    required IconData icon,
-    required String title,
-    required String description,
-  }) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: Row(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(10),
-            decoration: BoxDecoration(
-              color: Colors.amber[100],
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Icon(icon, color: Colors.amber[700], size: 26),
-          ),
-          const Gap(12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  title,
-                  style: const TextStyle(
-                    fontSize: 16,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                Text(
-                  description,
-                  style: const TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.w400,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _handlePurchase(L10n l10n) async {
+  Future<void> _handlePurchase() async {
+    final l10n = L10n.of(context);
     try {
-      final result = await ref
-          .read(
-            settingViewModelProvider().notifier,
-          )
-          .purchase();
+      final result =
+          await ref.read(settingViewModelProvider().notifier).purchase();
+      if (!context.mounted) {
+        return;
+      }
       if (result) {
         controller.play();
-        try {
-          await showDialog<void>(
-            context: context,
-            barrierColor: Colors.black87,
-            barrierDismissible: false,
-            builder: (_) {
-              return AlertDialog(
-                backgroundColor: Colors.transparent,
-                title: SizedBox(
-                  height: 550,
-                  child: Column(
-                    children: [
-                      Center(
-                        child: Container(
-                          width: 200,
-                          height: 0,
-                          alignment: Alignment.topCenter,
-                          child: ConfettiWidget(
-                            confettiController: controller,
-                            blastDirectionality: BlastDirectionality.explosive,
-                            emissionFrequency: 1,
-                            numberOfParticles: 30,
-                            maxBlastForce: 5,
-                            minBlastForce: 2,
-                            gravity: 0.3,
-                            createParticlePath: drawStar,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      Container(
-                        color: Colors.transparent,
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 4,
-                          vertical: 10,
-                        ),
-                        child: Column(
-                          children: [
-                            Text(
-                              l10n.paywallWelcomeTitle,
-                              style: PaywallStyle.wellComeTitle(),
-                              textAlign: TextAlign.center,
-                            ),
-                            const Gap(30),
-                            Assets.image.present.image(
-                              width: 200,
-                              height: 200,
-                            ),
-                          ],
-                        ),
-                      ),
-                      const Spacer(),
-                    ],
-                  ),
-                ),
-              );
-            },
-          ).timeout(
-            const Duration(seconds: 3),
-          );
-        } on TimeoutException {
-          if (context.mounted) {
-            context.pop();
-            await pageController.nextPage(
-              duration: const Duration(
-                milliseconds: 500,
-              ),
-              curve: Curves.easeInOut,
-            );
-          }
-        }
-      } else {
-        if (context.mounted) {
-          await pageController.nextPage(
-            duration: const Duration(
-              milliseconds: 500,
-            ),
-            curve: Curves.easeInOut,
-          );
-        }
+        await showPaywallSuccessDialog(
+          context: context,
+          controller: controller,
+        );
+        _goToNextPage();
       }
     } on Exception catch (e) {
       if (context.mounted) {
@@ -501,14 +261,16 @@ class _TutorialScreenState extends ConsumerState<TutorialScreen> {
             content: Text('${l10n.purchaseError}: $e'),
           ),
         );
-        await pageController.nextPage(
-          duration: const Duration(
-            milliseconds: 500,
-          ),
-          curve: Curves.easeInOut,
-        );
+        _goToNextPage();
       }
     }
+  }
+
+  void _goToNextPage() {
+    pageController.nextPage(
+      duration: const Duration(milliseconds: 500),
+      curve: Curves.easeInOut,
+    );
   }
 }
